Hs jwt topic

+++ src/main/java/shop/mtcoding/miniproject2/service/ResumeService.java
@@ -71,7 +71,11 @@ public class ResumeService {
 
     public void updateById(int id, int pInfoId, ResumeUpdateReqDto resumeUpdateReqDto) {
         String uuidImageName = PathUtil.writeImageFile(resumeUpdateReqDto.getProfile());
+        Resume resumePS = resumeRepository.findById(id);
 
+        if (resumePS.getPInfoId() != pInfoId) {
+            throw new CustomException("이력서 수정 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
         ResumeUpdateReqBirthdayTimestampDto resumeUpdateReqBirthdayTimestampDto = new ResumeUpdateReqBirthdayTimestampDto(
                 resumeUpdateReqDto.getTitle(),
                 resumeUpdateReqDto.getPortfolio(),
@@ -115,7 +119,13 @@ public class ResumeService {
 
     }
 
-    public void delete(int id) {
+    public void delete(int id, int pInfoId) {
+        Resume resumePS = resumeRepository.findById(id);
+
+        if (resumePS.getPInfoId() != pInfoId) {
+            throw new CustomApiException("이력서 삭제 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
+
         int result = resumeRepository.deleteById(id);
         if (result != 1) {
             throw new CustomApiException("이력서 삭제 실패하였습니다", HttpStatus.INTERNAL_SERVER_ERROR);
